Building a documentation assistant/requirements.txt
langchain
langchain-openai
langchain-community
langchain-aws
langchain-text-splitters
langgraph
openai
pinecone
langchain-pinecone
langchain-chroma
streamlit
python-dotenv
pypdf
python-docx
unstructured
beautifulsoup4
tiktoken
numpy
pandas
tqdm
tavily-python
langchain-tavily
boto3
opensearch-py
requests-aws4auth
